requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pytesseract  # (you can remove later if you want)
google-cloud-vision
google-auth
google-auth-oauthlib
google-auth-httplib2
